Make Adjacency fit the selected mode Internal use.

Make Adjacency fit the selected mode Internal use.

apply_mode(
  raw.scanList,
  mode = c("directed", "undirected", "max", "min", "upper", "lower", "plus", "vector")
)

Arguments

raw.scanList
a list of (directed by default) binary adjacency matrix, pre-theoretical, to which the chosen mode is to be applied

mode
Character scalar, specifies how igraph should interpret the supplied matrix. Default here is directed. Possible values are: directed, undirected, upper, lower, max, min, plus. Added vector too. See details graph_from_adjacency_matrix().

Value

an list of adjacency matrix fitting the chosen mode is to be applied