Count number of times edges have not been observed For empirical scanList, an edge that is not observed during a scan is NA. This function counts how many time this was the case in the inputted scan.list for each edge.

Written as a S3 method to be applied to scanList or sLlist (list of scanList) objects.

count_NA(scan.list, ...)

Arguments

scan.list
a scanList or sLlist object, where an unobserved edge (whether it is 0 or 1) is NA. See objects returned by simunet()

...
additional arguments to be passed. At the moment scale_scans() does not use

At the moment count_NA() does not use additional argument, arguments passed will be ignored.

Value

an integer matrix, or list of such, representing how many time each edge has been unobserved (i.e. was NA). Inherits from the previous scanList class (theoretical or empirical, inheriting from scanList), and keeps track of the scan.list's list of attributes attrs.

See also

Examples

set.seed(42)
n <- 5L
samp.effort <- 241L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
diag(Adj) <- 0L
Adj

#>     a   b   c   d   e
#> a   0 228  47 165 111
#> b 229   0  24 110 131
#> c  65 122   0  20  41
#> d 153 241 100   0 227
#> e  74 128  89 114   0

# social network simulations
## theoretical scans
sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "directed",n.scans = 120L)
sL

#> , , 1
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 0 0 1 0 0
#> 
#> , , 2
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 1 0
#> c 1 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 1 1 0
#> 
#> , , 3
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 1 0 1 0
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 4
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 0
#> c 1 1 0 1 0
#> d 1 1 1 0 1
#> e 1 1 1 0 0
#> 
#> , , 5
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 6
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 1 0 1 0 0
#> 
#> , , 7
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 8
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 9
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 1 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 10
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 0 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 11
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 1 0 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 12
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 0 0 0 1
#> d 1 1 0 0 1
#> e 1 1 1 0 0
#> 
#> , , 13
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 0
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 14
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 15
#> 
#>   a b c d e
#> a 0 0 0 1 0
#> b 1 0 0 1 1
#> c 1 0 0 1 0
#> d 1 1 1 0 1
#> e 1 0 1 0 0
#> 
#> , , 16
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 0
#> c 1 0 0 1 0
#> d 0 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 17
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 1 1 0 0 0
#> d 1 1 0 0 1
#> e 1 1 1 0 0
#> 
#> , , 18
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 1 0 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 19
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 1
#> c 1 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 20
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 21
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 0 0 0 1 0
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 22
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 1 1 0 1 0
#> 
#> , , 23
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 0
#> c 0 0 0 1 0
#> d 1 1 0 0 1
#> e 0 0 1 0 0
#> 
#> , , 24
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 25
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 0
#> e 1 0 0 0 0
#> 
#> , , 26
#> 
#>   a b c d e
#> a 0 1 1 1 1
#> b 1 0 0 1 1
#> c 0 1 0 1 0
#> d 0 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 27
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 0 0 0 1 1
#> c 1 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 28
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 1 0 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 29
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 1
#> c 0 1 0 0 1
#> d 0 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 30
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 31
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 32
#> 
#>   a b c d e
#> a 0 0 0 0 0
#> b 1 0 0 0 0
#> c 0 1 0 0 1
#> d 0 1 1 0 1
#> e 0 0 1 0 0
#> 
#> , , 33
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 1
#> c 1 0 0 0 1
#> d 0 1 0 0 0
#> e 0 1 0 0 0
#> 
#> , , 34
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 1 1 1 0 1
#> e 0 0 1 1 0
#> 
#> , , 35
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 1 0 0
#> 
#> , , 36
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 1 1 0 0 0
#> d 1 1 0 0 1
#> e 1 0 1 0 0
#> 
#> , , 37
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 1
#> c 0 1 0 0 1
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 38
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 1 0 0 0 0
#> d 0 1 0 0 1
#> e 1 1 0 0 0
#> 
#> , , 39
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 0
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 40
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 0
#> c 0 0 0 0 1
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 41
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 1 0 1
#> e 0 0 1 1 0
#> 
#> , , 42
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 43
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 1 0 1
#> c 1 0 0 0 1
#> d 0 1 0 0 1
#> e 1 0 1 0 0
#> 
#> , , 44
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 1 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 1 0
#> 
#> , , 45
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 46
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 47
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 48
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 0
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 49
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 1 0 1 1 0
#> 
#> , , 50
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 51
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 0 0 0 1 0
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 1 1 1 1 0
#> 
#> , , 52
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 53
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 1 0 1 0 0
#> 
#> , , 54
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 1 1 1 0 0
#> e 0 1 0 0 0
#> 
#> , , 55
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 56
#> 
#>   a b c d e
#> a 0 0 0 0 1
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 57
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 1 0
#> c 1 1 0 0 1
#> d 0 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 58
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 1 0
#> c 1 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 1 0
#> 
#> , , 59
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 1 0
#> c 1 1 0 0 0
#> d 1 1 1 0 1
#> e 0 0 0 0 0
#> 
#> , , 60
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 61
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 1 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 62
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 1 1
#> c 1 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 63
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 1 1 0
#> c 0 1 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 64
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 65
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 1
#> c 1 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 1 0
#> 
#> , , 66
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 1
#> c 0 0 0 0 1
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 67
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 68
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 1 0 1
#> c 0 0 0 0 1
#> d 1 1 1 0 1
#> e 1 1 0 1 0
#> 
#> , , 69
#> 
#>   a b c d e
#> a 0 0 0 1 1
#> b 1 0 1 1 0
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 70
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 1 0 0
#> 
#> , , 71
#> 
#>   a b c d e
#> a 0 1 1 1 1
#> b 1 0 1 0 0
#> c 0 0 0 1 0
#> d 1 1 0 0 1
#> e 1 1 0 0 0
#> 
#> , , 72
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 73
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 1 1 0 0 1
#> d 1 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 74
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 1
#> c 0 1 0 0 1
#> d 0 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 75
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 1
#> c 0 0 0 1 1
#> d 0 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 76
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 1 0 0
#> 
#> , , 77
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 1
#> c 0 0 0 1 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 78
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 0 0 1 1 1
#> c 1 1 0 0 0
#> d 0 1 0 0 1
#> e 1 1 1 0 0
#> 
#> , , 79
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 1 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 80
#> 
#>   a b c d e
#> a 0 1 1 1 1
#> b 0 0 0 1 1
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 81
#> 
#>   a b c d e
#> a 0 1 1 1 1
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 1 0 0 0 0
#> 
#> , , 82
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 1 1 0 1 0
#> 
#> , , 83
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 1 0 1 0 0
#> 
#> , , 84
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 1
#> c 1 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 1 0
#> 
#> , , 85
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 1 0 0
#> c 1 0 0 0 0
#> d 1 1 0 0 1
#> e 1 1 1 0 0
#> 
#> , , 86
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 1 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 0 0
#> 
#> , , 87
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 1 1 1 0 1
#> e 1 0 0 1 0
#> 
#> , , 88
#> 
#>   a b c d e
#> a 0 1 1 1 1
#> b 1 0 0 1 1
#> c 0 0 0 0 0
#> d 1 1 0 0 1
#> e 0 0 1 1 0
#> 
#> , , 89
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 90
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 0 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 1 1 0
#> 
#> , , 91
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 0 0
#> c 1 0 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> , , 92
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 1 0 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 1 0 0 0 0
#> 
#> , , 93
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 0
#> c 1 1 0 1 0
#> d 1 1 0 0 1
#> e 1 0 0 1 0
#> 
#> , , 94
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 1 0
#> 
#> , , 95
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 0 0
#> c 1 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 1 0 0
#> 
#> , , 96
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 1 1 0 0 1
#> d 1 1 0 0 1
#> e 1 0 0 1 0
#> 
#> , , 97
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 1 1
#> c 1 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 1 0 0
#> 
#> , , 98
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 0 1 1 1 0
#> 
#> , , 99
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 1 0 0
#> 
#> , , 100
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 0 1
#> c 1 0 0 0 0
#> d 0 1 0 0 1
#> e 0 0 1 0 0
#> 
#> , , 101
#> 
#>   a b c d e
#> a 0 0 0 1 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 102
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 0
#> c 0 1 0 0 1
#> d 0 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 103
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 0 0 0 0 1
#> c 1 1 0 0 0
#> d 0 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 104
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 0
#> c 0 0 0 0 0
#> d 0 1 0 0 1
#> e 1 0 0 1 0
#> 
#> , , 105
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 0 1
#> c 0 0 0 0 0
#> d 0 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 106
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 0 0
#> c 0 1 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 1 0
#> 
#> , , 107
#> 
#>   a b c d e
#> a 0 1 0 0 0
#> b 1 0 0 0 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 0 1 0
#> 
#> , , 108
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 1 1 1 1 0
#> 
#> , , 109
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 1 0 1
#> c 0 1 0 0 0
#> d 0 1 0 0 1
#> e 1 1 1 0 0
#> 
#> , , 110
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 1 1
#> c 1 0 0 0 0
#> d 1 1 1 0 1
#> e 0 1 0 0 0
#> 
#> , , 111
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 1 1
#> c 0 1 0 1 0
#> d 1 1 1 0 1
#> e 0 1 1 0 0
#> 
#> , , 112
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 1 0
#> c 0 0 0 1 1
#> d 0 1 0 0 0
#> e 1 1 1 1 0
#> 
#> , , 113
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 0 0 0 1 1
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 0 1 0
#> 
#> , , 114
#> 
#>   a b c d e
#> a 0 1 1 0 0
#> b 1 0 0 0 0
#> c 0 1 0 1 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 115
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 0 1
#> c 1 1 0 0 0
#> d 0 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 116
#> 
#>   a b c d e
#> a 0 1 1 0 1
#> b 1 0 0 1 0
#> c 0 0 0 0 0
#> d 1 1 1 0 1
#> e 0 0 1 1 0
#> 
#> , , 117
#> 
#>   a b c d e
#> a 0 1 0 0 1
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 0 1 1 0 1
#> e 0 1 1 1 0
#> 
#> , , 118
#> 
#>   a b c d e
#> a 0 1 0 1 0
#> b 1 0 0 1 1
#> c 1 0 0 1 1
#> d 0 1 1 0 0
#> e 1 1 1 0 0
#> 
#> , , 119
#> 
#>   a b c d e
#> a 0 1 1 1 0
#> b 1 0 0 1 1
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 0 0 0 0
#> 
#> , , 120
#> 
#>   a b c d e
#> a 0 1 0 1 1
#> b 1 0 0 1 0
#> c 0 1 0 0 0
#> d 1 1 0 0 1
#> e 0 1 1 0 0
#> 
#> attr(,"class")
#> [1] "theoretical" "scanList"   
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob

## group-scan sampling
sL |> perform_exp(design_exp(customize_sampling("group",.6))) |> count_NA()

#>    a  b  c  d  e
#> a  0 49 54 44 55
#> b 50  0 41 48 45
#> c 43 56  0 54 54
#> d 45 57 55  0 40
#> e 53 53 49 45  0