Generator for empirical scanList objects Internal use. The user should rather rely on simunet() and/or perform_exp() as a wrapper for the different steps needed to perform simulations and experimental manipulations.

Generator for empirical scanList objects Internal use. The user should rather rely on simunet() and/or perform_exp() as a wrapper for the different steps needed to perform simulations and experimental manipulations.

generate_empiscanList(scan.list, exp.design)

Arguments

scan.list
a scanList object (see simunet())

exp.design
an expDesign object. See objects returned by design_exp()

Value

an empirical inheriting from scanList object, primarily a 3 dimensional array representing the (binary) adjacency matrices (coded within the first two dimensions of the 3D-array) obtained at each simulated scan (coded as the 3rd dimension of the 3D-array), and a list of attributes, attrs.

The list of attributes attrs contains:

all the previous attributes contained in scan.list's attrs attributes list, as well as:

scanList.type: character scalar, changed from "theoretical" to "empirical"

theoretical.scanList: the 3D array before the experimental manipulations contained in the inputted exp.design have been applied

See also