Shortcut to a lapply equivalent to apply a function to each 2D matrix contained in a scanList Written analogously to vapply(). Values returned by .f should be a similarly dimensionned matrix as the first one contained in the 3D array

Shortcut to a lapply equivalent to apply a function to each 2D matrix contained in a scanList Written analogously to vapply(). Values returned by .f should be a similarly dimensionned matrix as the first one contained in the 3D array

sLapply(sL, FUN, ...)

Arguments

sL
a scanList object (see simunet())

FUN
a function,to apply a function to each 2D matrix contained in sL

...
extra argument to be passed, notably named arguments used by .f (see lapply())

Value

a 3D array onto which the function has been applied to each scan

Examples

set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

#>   a   b  c  d  e
#> a 0 100 37 27 58
#> b 0   0 20 36 42
#> c 0   0  0  5 93
#> d 0   0  0  0 30
#> e 0   0  0  0  0

sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "upper",n.scans = 120L)
sL |> sLapply(\(scan) {scan[1,2] <- NA;scan})

#> [[1]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[2]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[3]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[4]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[5]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[6]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[7]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[8]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[9]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[10]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[11]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[12]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[13]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[14]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[15]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[16]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[17]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[18]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[19]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[20]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[21]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[22]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[23]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[24]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[25]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[26]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 0 0
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[27]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[28]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[29]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[30]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[31]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[32]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[33]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[34]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[35]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[36]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 1 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[37]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[38]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[39]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[40]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[41]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[42]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[43]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[44]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[45]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[46]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[47]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[48]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[49]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[50]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[51]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 0
#> c 0  0 0 1 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[52]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[53]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[54]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[55]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[56]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[57]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[58]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[59]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 1 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[60]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[61]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[62]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[63]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[64]]
#>   a  b c d e
#> a 0 NA 1 1 0
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[65]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[66]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[67]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 0
#> c 0  0 0 1 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[68]]
#>   a  b c d e
#> a 0 NA 1 1 0
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[69]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[70]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[71]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[72]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[73]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[74]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[75]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[76]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[77]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[78]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[79]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[80]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 1 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[81]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[82]]
#>   a  b c d e
#> a 0 NA 1 1 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[83]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[84]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[85]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 1
#> c 0  0 0 1 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[86]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 1 1
#> c 0  0 0 1 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[87]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[88]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[89]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[90]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[91]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[92]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[93]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[94]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[95]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[96]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[97]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 1 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[98]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[99]]
#>   a  b c d e
#> a 0 NA 1 1 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[100]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[101]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[102]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[103]]
#>   a  b c d e
#> a 0 NA 1 1 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[104]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[105]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 1 0 0
#> c 0  0 0 1 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[106]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[107]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[108]]
#>   a  b c d e
#> a 0 NA 0 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[109]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 0 0
#> c 0  0 0 1 0
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[110]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[111]]
#>   a  b c d e
#> a 0 NA 1 0 0
#> b 0  0 0 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[112]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[113]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[114]]
#>   a  b c d e
#> a 0 NA 1 0 1
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#> 
#> [[115]]
#>   a  b c d e
#> a 0 NA 0 1 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[116]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[117]]
#>   a  b c d e
#> a 0 NA 1 1 0
#> b 0  0 0 0 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[118]]
#>   a  b c d e
#> a 0 NA 0 1 0
#> b 0  0 0 0 0
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[119]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 0 1 1
#> c 0  0 0 0 1
#> d 0  0 0 0 0
#> e 0  0 0 0 0
#> 
#> [[120]]
#>   a  b c d e
#> a 0 NA 0 0 1
#> b 0  0 1 1 0
#> c 0  0 0 0 1
#> d 0  0 0 0 1
#> e 0  0 0 0 0
#>