Scale list of binary scans into a weighted adjacency matrix Scaling here is dividing by the sum of 1s by the number of time an edge has been observed (whether it was 0 or 1).

Written as a S3 method to be applied to scanList or sLlist (list of scanList) objects.

scale_scans(scan.list, ...)

Arguments

scan.list
a scanList or sLlist object. See objects returned by simunet()

...
additional arguments to be passed. At the moment scale_scans() does not use

At the moment scale_scans() does not use additional argument, arguments passed will be ignored.

Value

a scaled object, or list of such, consisting mainly on a weighted adjacency matrix where each edge weight is equal to the sum of all binary edges divided by the number of times they have been sampled (determined via count_nonNA()). Inherits from weightedAdj and the previous scanList class (theoretical or empirical, inheriting from scanList), and keeps track of the scan.list's list of attributes attrs

See also

Examples

set.seed(42)
n <- 5L
samp.effort <- 241L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
diag(Adj) <- 0L
Adj

#>     a   b   c   d   e
#> a   0 228  47 165 111
#> b 229   0  24 110 131
#> c  65 122   0  20  41
#> d 153 241 100   0 227
#> e  74 128  89 114   0

# social network simulations
## theoretical scans
sL <- simunet(Adj = Adj,samp.effort = samp.effort,mode = "directed",n.scans = 120L)
sL

#> 
#> scan: 1           
#> a . 1 . 1 .
#> b 1 . . . .
#> c . . . . .
#> d . 1 1 . 1
#> e . . 1 . .
#> 
#> scan: 2           
#> a . 1 . . .
#> b 1 . . 1 .
#> c 1 1 . . .
#> d . 1 . . 1
#> e . . 1 1 .
#> 
#> ... ( 117  more scans)
#> 
#> scan: 120           
#> a . 1 . 1 1
#> b 1 . . 1 .
#> c . 1 . . .
#> d 1 1 . . 1
#> e . 1 1 . .
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob

# scale_scans() can scale `weightedAdj` objects...
sL |> sum_scans() |> scale_scans()

#> 
#> scan: Weighted adjacency matrix                          
#> a .    0.96 0.30 0.52 0.40
#> b 0.93 .    0.09 0.40 0.53
#> c 0.30 0.50 .    0.12 0.14
#> d 0.59 1.00 0.38 .    0.96
#> e 0.24 0.50 0.39 0.45 .   
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob summed.scanList sampled

# ... or `scanList` object directly
## group-scan sampling
sL |> perform_exp(design_exp(customize_sampling("group",.6))) |> scale_scans()

#> 
#> scan: Weighted adjacency matrix                          
#> a .    0.94 0.26 0.49 0.35
#> b 0.93 .    0.06 0.35 0.49
#> c 0.34 0.42 .    0.12 0.14
#> d 0.67 1.00 0.34 .    0.95
#> e 0.22 0.51 0.41 0.47 .   
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob obs.P theoretical.scanList summed.scanList sampled

## comparing group-scan and focal sampling
sL |> perform_exp(design_exp(customize_sampling("group",.6)),
                  design_exp(customize_sampling("focal","even"))
                  ) |> scale_scans()

#> [[1]]
#> 
#> scan: Weighted adjacency matrix                          
#> a .    0.97 0.39 0.56 0.49
#> b 0.92 .    0.07 0.41 0.46
#> c 0.29 0.43 .    0.11 0.18
#> d 0.58 1.00 0.38 .    0.96
#> e 0.25 0.47 0.39 0.45 .   
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob obs.P theoretical.scanList summed.scanList sampled
#> [[2]]
#> 
#> scan: Weighted adjacency matrix                          
#> a .    0.96 0.35 0.65 0.42
#> b 0.88 .    0.10 0.38 0.44
#> c 0.31 0.52 .    0.15 0.08
#> d 0.60 1.00 0.40 .    1.00
#> e 0.19 0.58 0.33 0.44 .   
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob focalList theoretical.scanList summed.scanList sampled
#> attr(,"class")
#> [1] "sLlist"