Convert scanList 3D array into list of 2D matrices

Convenience function to allow the use of functions designed to affect list of matrices instead of 3D array (specifically scanList objects) where the 3rd dimension is the scan index. Keeps track of the attributes list attrs

scanList2matList(scan.list)

Arguments

scan.list
a 3 dimensional array representing adjacency matrices (first 2 dimensions) throughout the different scans (3rd dimension)

Value

a list of matrices where each element is the 2D matrix for each scan index (3rd dimension of array.3D)

Details

To back-transform a list of matrices into a scanList object, users can use matList2array(), wrapper to simplify2array() that handles attributes list attrs

Examples

set.seed(42)
n <- 5L
samp.effort <- 100L

# Adjacency matrix import
## random directed adjacency matrix
Adj <- sample(1:samp.effort,n * n) |>
  matrix(nrow = 5,dimnames = list(letters[1:n],letters[1:n]))
Adj[lower.tri(Adj,diag = TRUE)] <- 0L
Adj

#>   a   b  c  d  e
#> a 0 100 37 27 58
#> b 0   0 20 36 42
#> c 0   0  0  5 93
#> d 0   0  0  0 30
#> e 0   0  0  0  0

mL <- simunet(Adj,samp.effort,"upper",10) |> scanList2matList()
mL

#> [[1]]
#>   a b c d e
#> a 0 1 1 0 1
#> b 0 0 0 0 0
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[2]]
#>   a b c d e
#> a 0 1 0 0 0
#> b 0 0 0 0 1
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[3]]
#>   a b c d e
#> a 0 1 0 0 0
#> b 0 0 0 0 1
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[4]]
#>   a b c d e
#> a 0 1 1 0 1
#> b 0 0 0 0 1
#> c 0 0 0 0 0
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[5]]
#>   a b c d e
#> a 0 1 0 0 1
#> b 0 0 0 1 0
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[6]]
#>   a b c d e
#> a 0 1 1 0 1
#> b 0 0 1 0 1
#> c 0 0 0 0 1
#> d 0 0 0 0 1
#> e 0 0 0 0 0
#> 
#> [[7]]
#>   a b c d e
#> a 0 1 0 1 0
#> b 0 0 0 1 1
#> c 0 0 0 0 1
#> d 0 0 0 0 1
#> e 0 0 0 0 0
#> 
#> [[8]]
#>   a b c d e
#> a 0 1 1 0 0
#> b 0 0 0 0 1
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[9]]
#>   a b c d e
#> a 0 1 0 0 1
#> b 0 0 0 0 1
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> [[10]]
#>   a b c d e
#> a 0 1 0 1 1
#> b 0 0 0 0 0
#> c 0 0 0 0 1
#> d 0 0 0 0 0
#> e 0 0 0 0 0
#> 
#> attr(,"class")
#> [1] "theoretical" "matList"    
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob

mL |> matList2scanList()

#> 
#> scan: 1           
#> a . 1 1 . 1
#> b . . . . .
#> c . . . . 1
#> d . . . . .
#> e . . . . .
#> 
#> scan: 2           
#> a . 1 . . .
#> b . . . . 1
#> c . . . . 1
#> d . . . . .
#> e . . . . .
#> 
#> ... ( 7  more scans)
#> 
#> scan: 10           
#> a . 1 . 1 1
#> b . . . . .
#> c . . . . 1
#> d . . . . .
#> e . . . . .
#> 
#> 
#> Hidden attributes: scanList.type raw.scanList Adj samp.effort n.scans mode Adj.subfun edge.Prob